Move tmp_h5py_file to conftest
